phase-3/backend/mcp-server/requirements.txt
mcp
pydantic
pydantic-settings
sqlmodel
psycopg2-binary
python-dotenv
pytest
pytest-asyncio
